src/main/java/com/emirhan/ingcasestudy/controller/AuthenticationController.java
package com.emirhan.ingcasestudy.controller;

import com.emirhan.ingcasestudy.dto.AuthenticationRequest;
import com.emirhan.ingcasestudy.dto.AuthenticationResponse;
import com.emirhan.ingcasestudy.dto.RegisterRequest;
import com.emirhan.ingcasestudy.service.AuthenticationService;
import io.swagger.v3.oas.annotations.Operation;
import io.swagger.v3.oas.annotations.tags.Tag;
import lombok.RequiredArgsConstructor;
import org.springframework.http.ResponseEntity;
import org.springframework.web.bind.annotation.PostMapping;
import org.springframework.web.bind.annotation.RequestBody;
import org.springframework.web.bind.annotation.RequestMapping;
import org.springframework.web.bind.annotation.RestController;

@RestController
@RequestMapping("/auth")
@RequiredArgsConstructor
public class AuthenticationController {

    private final AuthenticationService authenticationService;

    @PostMapping("/register")
    @Operation(summary = "Register User", description = "This endpoint register user and returns a token.")
    public ResponseEntity<AuthenticationResponse> register(@RequestBody RegisterRequest request) {
        return ResponseEntity.ok(authenticationService.register(request));
    }


    @PostMapping("/authenticate")
    @Operation(summary = "Authenticate User", description = "This endpoint authenticate user and returns a token.")
    public ResponseEntity<AuthenticationResponse> authenticate(@RequestBody AuthenticationRequest request) {
        return ResponseEntity.ok(authenticationService.authenticate(request));
    }
}
